+++ apps/web/src/components/Messages/Push/MessageHeader.tsx
@@ -1,13 +1,13 @@
 import Follow from '@components/Shared/Follow';
 import Unfollow from '@components/Shared/Unfollow';
 import UserProfile from '@components/Shared/UserProfile';
-import useGroupInfoModal from '@components/utils/hooks/push/usePushGroupInfo';
+import useGroupInfoModal from '@components/utils/hooks/push/useGroupInfoModal';
 import useOnClickOutside from '@components/utils/hooks/useOnClickOutside';
 import type { GroupDTO, IFeeds } from '@pushprotocol/restapi';
 import type { Profile } from 'lens';
 import React, { useEffect, useRef, useState } from 'react';
 import { CHAT_TYPES, usePushChatStore } from 'src/store/push-chat';
-import { Image } from 'ui';
+import { Image, Modal } from 'ui';
 
 interface MessageHeaderProps {
   profile?: Profile;
@@ -28,11 +28,18 @@ export default function MessageHeader({
   const selectedChatType = usePushChatStore((state) => state.selectedChatType);
   const lensProfiles = usePushChatStore((state) => state.lensProfiles);
   const [showModal, setShowModal] = useState(false);
-  const [showGroupInfoModal, setShowGroupInfoModal] = useState(false);
+  const showGroupInfoModal = usePushChatStore(
+    (state) => state.showGroupInfoModal
+  );
+  const setShowGroupInfoModal = usePushChatStore(
+    (state) => state.setShowGroupInfoModal
+  );
+
+  // const [showGroupInfoModal, setShowGroupInfoModal] = useState(false);
 
   const downRef = useRef(null);
   useOnClickOutside(downRef, () => {
-    setShowGroupInfoModal(false);
+    // setShowGroupInfoModal(false);
     setShowModal(false);
   });
 
@@ -43,6 +50,26 @@ export default function MessageHeader({
     const profile = lensProfiles.get(selectedChatId);
     setFollowing(profile?.isFollowedByMe ?? false);
   }, [lensProfiles, selectedChatId, selectedChatType]);
+
+  const {
+    groupInfoModal,
+    modalContent: groupInfoModalContent,
+    isModalClosable: isGroupInfoModalClosable
+  } = useGroupInfoModal({
+    groupInfo: groupInfo,
+    setGroupInfo: setGroupInfo
+    // show: showGroupInfoModal,
+    // setShow: setShowGroupInfoModal
+  });
+
+  const handleGroupInfo = async () => {
+    try {
+      groupInfoModal();
+      setShowModal(false);
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   return (
     <section className="flex w-full justify-between border-b px-5	py-2.5">
@@ -63,7 +90,11 @@ export default function MessageHeader({
         )}
       </div>
       <div className="flex items-center gap-4	">
-        <img className="cursor-pointer" src="/push/video.svg" alt="video icon" />
+        <img
+          className="cursor-pointer"
+          src="/push/video.svg"
+          alt="video icon"
+        />
         {profile &&
           (following ? (
             <Unfollow profile={profile!} setFollowing={setFollowing} showText />
@@ -73,32 +104,52 @@ export default function MessageHeader({
         {groupInfo && (
           <div
             className="w-fit cursor-pointer"
-            onClick={() => (showModal === false ? setShowModal(true) : setShowModal(false))}
+            onClick={() =>
+              showModal === false ? setShowModal(true) : setShowModal(false)
+            }
           >
-            <Image className="h-10 w-9" src="/push/more.svg" alt="group info settings" />
+            <Image
+              className="h-10 w-9"
+              src="/push/more.svg"
+              alt="group info settings"
+            />
           </div>
         )}
         {groupInfo && showModal && (
           <div
             ref={downRef}
             className="absolute top-36 ml-[-80px] flex w-40 cursor-pointer items-center rounded-2xl border border-[#BAC4D6] bg-white p-2 px-4"
-            onClick={() => {
-              setShowGroupInfoModal(true);
-              setShowModal(false);
-            }}
+            onClick={handleGroupInfo}
           >
             <div>
-              <Image className="mr-2 h-6 w-6" src="/push/info.svg" alt="group info settings" />
+              <Image
+                className="mr-2 h-6 w-6"
+                src="/push/info.svg"
+                alt="group info settings"
+              />
             </div>
-            <div className="items-center text-[17px] font-[400] text-[#657795]">Group Info</div>
+            <div className="items-center text-[17px] font-[400] text-[#657795]">
+              Group Info
+            </div>
           </div>
         )}
-        {useGroupInfoModal({
+        {/* {(useGroupInfoModal({
           groupInfo: groupInfo,
           setGroupInfo: setGroupInfo,
           show: showGroupInfoModal,
           setShow: setShowGroupInfoModal
-        })}
+        }))} */}
+        <Modal
+          size="xs"
+          show={showGroupInfoModal}
+          onClose={
+            isGroupInfoModalClosable
+              ? () => setShowGroupInfoModal(false)
+              : () => {}
+          }
+        >
+          {groupInfoModalContent}
+        </Modal>
       </div>
     </section>
   );
