+++ apps/web/src/components/utils/hooks/push/index.ts
@@ -10,9 +10,9 @@ export * from './useGetChatProfile';
 export * from './useGetGroup';
 export * from './useGetGroupbyName';
 export * from './useGetUsersBatch';
+export * from './useGroupInfoModal';
 export * from './usePushChatSocket';
 export * from './usePushCreateGroupChat';
 export * from './usePushDecryption';
-export * from './usePushGroupInfo';
 export * from './usePushSendMessage';
 export * from './useUpgradeChatProfile';
