+++ apps/web/src/components/utils/hooks/push/usePushCreateGroupChat.tsx
@@ -84,7 +84,8 @@ export const MemberProfileList = ({
 }: memberProfileListType) => {
   // have used this state instead of boolean since if I used boolean than I still had to check if the user id is equal to the id selected to make it unique and open only that modal which is suppose to open and not open every modal
   const [showModalgroupOptions, setShowModalgroupOptions] = useState(-1);
-  const [showModalAdmingroupOptions, setShowModalAdmingroupOptions] = useState(-2);
+  const [showModalAdmingroupOptions, setShowModalAdmingroupOptions] =
+    useState(-2);
 
   const handleCloseAllmodals = () => {
     setShowModalgroupOptions(-1);
@@ -93,7 +94,9 @@ export const MemberProfileList = ({
   const downRef = useRef(null);
   useOnClickOutside(downRef, () => handleCloseAllmodals());
   const currentProfile = useAppStore((s) => s.currentProfile);
-  const setShowCreateGroupModal = usePushChatStore((s) => s.setShowCreateGroupModal);
+  const setShowCreateGroupModal = usePushChatStore(
+    (s) => s.setShowCreateGroupModal
+  );
 
   const handleRemoveClick = (profile: Profile) => {
     if (onRemoveMembers) {
@@ -177,9 +180,15 @@ export const MemberProfileList = ({
 
             <div className="flex flex-col">
               {!isAddedMembersList && isAdmin(member) && (
-                <p className="truncate font-bold">{member?.name ?? formatHandle(member?.handle)}</p>
+                <p className="truncate font-bold">
+                  {member?.name ?? formatHandle(member?.handle)}
+                </p>
               )}
-              <Slug className="text-sm" slug={formatHandle(member?.handle)} prefix="@" />
+              <Slug
+                className="text-sm"
+                slug={formatHandle(member?.handle)}
+                prefix="@"
+              />
             </div>
           </div>
           {isAdmin(member) && !onAddMembers ? (
@@ -193,9 +202,15 @@ export const MemberProfileList = ({
             <div className="relative flex">
               <div
                 className="w-fit cursor-pointer"
-                onClick={() => setShowModalgroupOptions(showModalgroupOptions === i ? -1 : i)}
+                onClick={() =>
+                  setShowModalgroupOptions(showModalgroupOptions === i ? -1 : i)
+                }
               >
-                <img className="h-10 w-9" src="/push/more.svg" alt="more icon" />
+                <img
+                  className="h-10 w-9"
+                  src="/push/more.svg"
+                  alt="more icon"
+                />
               </div>
               {showModalgroupOptions === i && (
                 <div
@@ -216,25 +231,43 @@ export const MemberProfileList = ({
                   </div>
                   <div
                     className="flex cursor-pointer items-center p-[8px] text-lg font-medium"
-                    onClick={() => (isAdmin(member) ? handleRemoveAdmin(member) : handleMakeAdmin(member))}
+                    onClick={() =>
+                      isAdmin(member)
+                        ? handleRemoveAdmin(member)
+                        : handleMakeAdmin(member)
+                    }
                   >
                     <Image
-                      src={isAdmin(member) ? '/push/dismissadmin.svg' : '/push/Shield.svg'}
+                      src={
+                        isAdmin(member)
+                          ? '/push/dismissadmin.svg'
+                          : '/push/Shield.svg'
+                      }
                       className="h-[20px] pr-[10px]"
                       alt="admin icon"
                     />
                     <div className="text-[15px] font-[450]">
-                      {isAdmin(member) ? `Dismiss as admin` : `Make group admin`}
+                      {isAdmin(member)
+                        ? `Dismiss as admin`
+                        : `Make group admin`}
                     </div>
                   </div>
                   <div
                     className="flex cursor-pointer items-center p-[8px] text-lg font-medium"
                     onClick={() =>
-                      isAdmin(member) ? handleRemoveUseradmin(member) : handleRemoveClick(member)
+                      isAdmin(member)
+                        ? handleRemoveUseradmin(member)
+                        : handleRemoveClick(member)
                     }
                   >
-                    <Image src="/push/MinusCircle.svg" className="h-[20px] pr-[10px]" alt="remove icon" />
-                    <div className="text-[15px] font-[450] text-red-600">Remove</div>
+                    <Image
+                      src="/push/MinusCircle.svg"
+                      className="h-[20px] pr-[10px]"
+                      alt="remove icon"
+                    />
+                    <div className="text-[15px] font-[450] text-red-600">
+                      Remove
+                    </div>
                   </div>
                 </div>
               )}
@@ -246,7 +279,8 @@ export const MemberProfileList = ({
               className="cursor-pointer rounded-lg border border-violet-600 px-2 text-violet-600"
               onClick={() => onAddMembers(member)}
             >
-              <span className="text-sm">Add</span> <span className="text-base">+</span>
+              <span className="text-sm">Add</span>{' '}
+              <span className="text-base">+</span>
             </div>
           )}
         </div>
@@ -259,7 +293,9 @@ const useCreateGroup = () => {
   const pgpPrivateKey = usePushChatStore((state) => state.pgpPrivateKey);
   const decryptedPgpPvtKey = pgpPrivateKey.decrypted;
   const currentProfile = useAppStore((state) => state.currentProfile);
-  const setShowCreateGroupModal = usePushChatStore((state) => state.setShowCreateGroupModal);
+  const setShowCreateGroupModal = usePushChatStore(
+    (state) => state.setShowCreateGroupModal
+  );
   const [step, setStep] = useState<number>(1);
   const [modalClosable, setModalClosable] = useState<boolean>(true);
   const [groupName, setGroupName] = useState<string>('');
@@ -367,7 +403,8 @@ const useCreateGroup = () => {
               messages: Array.isArray(chats.get(response.chatId)?.messages)
                 ? [...chats.get(response.chatId)!.messages, msg]
                 : [msg],
-              lastThreadHash: chats.get(response.chatId)?.lastThreadHash ?? msg.link
+              lastThreadHash:
+                chats.get(response.chatId)?.lastThreadHash ?? msg.link
             });
           }
         }
@@ -396,7 +433,10 @@ const useCreateGroup = () => {
     if (!e.target.files) {
       return;
     }
-    if ((e.target as HTMLInputElement).files && ((e.target as HTMLInputElement).files as FileList).length) {
+    if (
+      (e.target as HTMLInputElement).files &&
+      ((e.target as HTMLInputElement).files as FileList).length
+    ) {
       var reader = new FileReader();
       reader.readAsDataURL(e.target.files[0]);
 
@@ -502,12 +542,16 @@ const useCreateGroup = () => {
     // Update the state with the updated members list
     setMembers(updatedMembers);
     // Remove the user from the `adminAddresses` array
-    setAdminAddresses(adminAddresses.filter((admin) => admin.id !== profile.id));
+    setAdminAddresses(
+      adminAddresses.filter((admin) => admin.id !== profile.id)
+    );
   };
 
   const removeUserAdmin = (profile: Profile) => {
     // Remove the user from the `adminAddresses` array
-    const updatedAdminAddresses = adminAddresses.filter((admin) => admin.id !== profile.id);
+    const updatedAdminAddresses = adminAddresses.filter(
+      (admin) => admin.id !== profile.id
+    );
     setAdminAddresses(updatedAdminAddresses);
 
     // Update the member's `isAdmin` property to false and remove from the `members` array
@@ -538,17 +582,30 @@ const useCreateGroup = () => {
           >
             <XIcon className="h-5 w-5" />
           </button>
-          <div className="mb-4 mt-1  text-center text-xl font-medium">{modalInfo.title}</div>
+          <div className="mb-4 mt-1  text-center text-xl font-medium">
+            {modalInfo.title}
+          </div>
 
-          <div onClick={handleUpload} className=" w-fit cursor-pointer self-center">
+          <div
+            onClick={handleUpload}
+            className=" w-fit cursor-pointer self-center"
+          >
             {!!!groupImage && (
               <div className="my-4 w-fit cursor-pointer rounded-[2.5rem] bg-gray-100 p-10">
-                <img className="h-11 w-11" src="/push/uploadImage.svg" alt="plus icon" />
+                <img
+                  className="h-11 w-11"
+                  src="/push/uploadImage.svg"
+                  alt="plus icon"
+                />
               </div>
             )}
             {!!groupImage && (
               <div className="my-4 h-28 w-28 cursor-pointer overflow-hidden rounded-3xl">
-                <img className="h-full w-full" src={groupImage} alt="group image" />
+                <img
+                  className="h-full w-full"
+                  src={groupImage}
+                  alt="group image"
+                />
               </div>
             )}
             <input
@@ -563,7 +620,9 @@ const useCreateGroup = () => {
           <div className="my-4">
             <div className="flex items-center justify-between">
               <div className="pb-2 text-base font-medium">Group Name</div>
-              <span className="text-sm text-slate-500">{50 - groupName.length}</span>
+              <span className="text-sm text-slate-500">
+                {50 - groupName.length}
+              </span>
             </div>
             <Input
               type="text"
@@ -575,15 +634,21 @@ const useCreateGroup = () => {
           </div>
           <div className="my-4">
             <div className="flex items-center justify-between">
-              <div className="pb-2 text-base font-medium">Group Description</div>
-              <span className="text-sm text-slate-500">{150 - groupDescription.length}</span>
+              <div className="pb-2 text-base font-medium">
+                Group Description
+              </div>
+              <span className="text-sm text-slate-500">
+                {150 - groupDescription.length}
+              </span>
             </div>
             <Input
               type="text"
               className="px-4 py-4 text-sm"
               value={groupDescription}
               autoComplete="off"
-              onChange={(e) => setGroupDescription(e.target.value.slice(0, 150))}
+              onChange={(e) =>
+                setGroupDescription(e.target.value.slice(0, 150))
+              }
             />
           </div>
           <div className="my-4 flex flex-row justify-center">
@@ -598,7 +663,9 @@ const useCreateGroup = () => {
                 onClick={() => setGroupType(option)}
               >
                 <p className="text-base font-medium">{option?.title}</p>
-                <p className="text-center text-xs font-thin text-gray-400">{option?.subTitle}</p>
+                <p className="text-center text-xs font-thin text-gray-400">
+                  {option?.subTitle}
+                </p>
               </div>
             ))}
           </div>
@@ -630,7 +697,9 @@ const useCreateGroup = () => {
           >
             <XIcon className="h-5 w-5" />
           </button>
-          <div className="mb-4 mt-1 text-center text-xl font-medium">{modalInfo.title}</div>
+          <div className="mb-4 mt-1 text-center text-xl font-medium">
+            {modalInfo.title}
+          </div>
           <div className="flex flex-row justify-between pt-4 text-base">
             <span className="font-medium">Add users</span>
             <span className="text-sm text-slate-500">{`0${
@@ -677,7 +746,9 @@ const useCreateGroup = () => {
             disabled={isModalInputsEmpty()}
             onClick={handleNext}
           >
-            <span className="px-8">{modalClosable ? 'Create Group' : 'Creating Group...'}</span>
+            <span className="px-8">
+              {modalClosable ? 'Create Group' : 'Creating Group...'}
+            </span>
           </Button>
         </div>
       );
@@ -707,7 +778,9 @@ const useCreateGroup = () => {
           >
             <XIcon className="h-5 w-5" />
           </button>
-          <div className="pb-4 text-center text-base font-medium">{modalInfo.title}</div>
+          <div className="pb-4 text-center text-base font-medium">
+            {modalInfo.title}
+          </div>
         </div>
       );
   }
